Renamed PackageNames to comply with java code conventions
