+++ src/main/java/com/travel_gates_mod/travel_gates/blocks/QuickGateItem.java
@@ -1,6 +1,6 @@
-package com.TravelGatesMod.TravelGates.blocks;
+package com.travel_gates_mod.travel_gates.blocks;
 
-import com.TravelGatesMod.TravelGates.travelgates;
+import com.travel_gates_mod.travel_gates.travelgates;
 import net.minecraft.block.Block;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.Item;
