+++ src/main/java/com/travel_gates_mod/travel_gates/util/network/server/UpdateGatePacket.java
@@ -1,7 +1,7 @@
-package com.TravelGatesMod.TravelGates.util.Network.Server;
+package com.travel_gates_mod.travel_gates.util.network.server;
 
-import com.TravelGatesMod.TravelGates.util.GateInfo;
-import com.TravelGatesMod.TravelGates.util.GateInfoHandler;
+import com.travel_gates_mod.travel_gates.util.GateInfo;
+import com.travel_gates_mod.travel_gates.util.GateInfoHandler;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.network.PacketBuffer;
 import net.minecraftforge.fml.network.NetworkEvent;
